requirements.txt
Django==3.2.4
djangorestframework==3.12.4
psycopg2==2.8.6
Pillow==8.1.2
python-decouple==3.4
gunicorn==20.0.4
PyJWT==2.1.0
django-allauth==0.44.0
django-cleanup==5.2.0